ref: refactor gparser
